requirements.txt
Fabric==1.11.1
Pygments==2.1.3
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.23
blessings==1.6
bpython==0.15
celery==3.1.23
curtsies==0.2.6
dogpile.cache==0.5.7
dogpile.core==0.4.1
ecdsa==0.13
greenlet==0.4.9
-e git+https://github.com/bwhite/hadoopy@ff39b4e6d4e6efaf1f571cf0f2c0e0d7ab28c2d6#egg=hadoopy-master
kombu==3.0.35
nose==1.3.7
numpy==1.11.0
paramiko==1.16.0
psutil==4.1.0
pycrypto==2.6.1
pytz==2016.3
requests==2.9.1
simplejson==3.8.2
six==1.10.0
wcwidth==0.1.6
wsgiref==0.1.2

